refactor(shared): split agents/skills.ts into one file per skill
Moves the 23-entry SKILLS registry from a single 322-line file into
`packages/shared/src/agents/skills/`, one skill per file plus an
`index.ts` aggregator and a `types.ts` for the shared SystemSkill type.
Public API (`SKILLS`, `resolveSkillMention`) is unchanged; the two
internal consumers (`agents/index.ts`, `agents/mcpProjection.ts`)
re-target the new `./skills/index.js` path.

Motivation: as we add more per-LV and per-Fraktion skills, editing a
single growing file gets harder to review and conflicts more often.
Per-file layout means a new skill is a one-file diff and each file is
small enough to scan at a glance.

Behavior is unchanged. Verified by tsc on @gruenerator/shared,
@gruenerator/chat, @gruenerator/api, @gruenerator/web,
@gruenerator/mcp — all green.
