--- packages/shared/src/agents/skills.ts
@@ -1,329 +1,0 @@
-import type { SystemAgentId } from './system.js';
-import type { Skill } from './types.js';
-
-type SystemSkill = Skill & { identifier: SystemAgentId };
-
-export const SKILLS = [
-  {
-    identifier: 'gruenerator-antrag',
-    title: 'Antrag',
-    description: 'Anträge & Anfragen',
-    iconKey: 'PiFileText',
-    avatar: '📝',
-    backgroundColor: '#316049',
-    mention: 'antrag',
-    skillCategory: 'dokumente',
-    isSystemDefault: true,
-    promptTemplate: 'Schreibe einen Antrag zum Thema: ',
-  },
-  {
-    identifier: 'gruenerator-antrag',
-    title: 'Kommunalpolitik',
-    description: 'Haushalt bewerten, beraten, Anträge & Resolutionen entwerfen',
-    iconKey: 'PiBuildings',
-    avatar: '🏛️',
-    backgroundColor: '#316049',
-    mention: 'kommunal',
-    skillCategory: 'dokumente',
-    isSystemDefault: true,
-  },
-  {
-    identifier: 'gruenerator-buergerservice',
-    title: 'Bürger*innenanfragen',
-    description: 'Bürger*innenanfragen beantworten',
-    iconKey: 'PiChatCircle',
-    avatar: '💬',
-    backgroundColor: '#316049',
-    mention: 'bürgerservice',
-    skillCategory: 'presse',
-    promptTemplate: 'Beantworte folgende Anfrage: ',
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit',
-    title: 'Pressemitteilung',
-    description: 'Pressemitteilungen verfassen',
-    iconKey: 'PiNewspaper',
-    avatar: '📰',
-    backgroundColor: '#316049',
-    mention: 'presse',
-    contextPrefix: '[Plattform: Pressemitteilung]',
-    skillCategory: 'presse',
-    isSystemDefault: true,
-    promptTemplate: 'Schreibe eine PM zum Thema: ',
-    skillSystemPrompt: `**PRESSEMITTEILUNG (ca. 2000 Zeichen):**
-
-Struktur:
-1. **Aussagekräftiger Titel** – klar und informativ
-2. **Lead-Absatz** – wichtigste W-Fragen (Wer, Was, Wann, Wo, Warum)
-3. **Hauptteil** – Details und Kontext
-4. **Zitat** – von der*dem angegebenen Zitatgeber*in, falls vorhanden
-5. **Hintergrundinformationen** – für journalistische Einordnung
-
-Stil: Journalistischer Nachrichtenstil, sachlich und objektiv, keine Emojis, aktive Sprache.
-
-**Beispiel-Suche:** Nutze IMMER \`gruenerator_pressemitteilung_examples\`, um echte PMs aus Landesverbänden als Vorlage zu finden. Orientiere dich an Aufbau, Lead-Absatz, Zitat-Setzung und Tonalität.`,
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit',
-    title: 'Instagram',
-    description: 'Instagram-Posts & Captions',
-    iconKey: 'PiInstagramLogo',
-    avatar: '📸',
-    backgroundColor: '#316049',
-    mention: 'instagram',
-    contextPrefix: '[Plattform: Instagram]',
-    skillCategory: 'social',
-    isSystemDefault: true,
-    promptTemplate: 'Post zu folgendem Thema: ',
-    skillSystemPrompt: `**INSTAGRAM-POST (max. 600 Zeichen):**
-
-Visuell, Emojis am Satzanfang/-ende für Barrierefreiheit, strategische Hashtags. Erste Zeile als Hook.
-
-**Beispiel-Suche:** Nutze IMMER \`search_examples\` mit \`platform="instagram"\`, um echte, erfolgreiche Beispiel-Posts zu finden. Orientiere dich an Ton, Aufbau und Formatierung.`,
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit',
-    title: 'Facebook',
-    description: 'Facebook-Posts & Beiträge',
-    iconKey: 'PiFacebookLogo',
-    avatar: '👍',
-    backgroundColor: '#316049',
-    mention: 'facebook',
-    contextPrefix: '[Plattform: Facebook]',
-    skillCategory: 'social',
-    promptTemplate: 'Beitrag zu folgendem Thema: ',
-    skillSystemPrompt: `**FACEBOOK-POST (max. 600 Zeichen):**
-
-Locker, gesprächig, Emojis sparsam, Community-fokussiert, klarer Call-to-Action am Ende.
-
-**Beispiel-Suche:** Nutze IMMER \`search_examples\` mit \`platform="facebook"\`, um echte, erfolgreiche Beispiel-Posts zu finden. Orientiere dich an Ton, Aufbau und Formatierung.`,
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit',
-    title: 'Twitter / X',
-    description: 'Tweets & Threads',
-    iconKey: 'PiXLogo',
-    avatar: '🐦',
-    backgroundColor: '#316049',
-    mention: 'twitter',
-    contextPrefix: '[Plattform: Twitter]',
-    skillCategory: 'social',
-    promptTemplate: 'Tweet zu folgendem Thema: ',
-    skillSystemPrompt: `**TWEET / X-POST (max. 280 Zeichen):**
-
-Prägnant, pointiert, direkte Sprache, sparsame Emojis. Eine zentrale Aussage statt mehrerer.
-
-**Beispiel-Suche:** Nutze IMMER \`search_examples\` mit \`platform="bluesky"\` (NICHT "twitter"/"x") — unsere Tweet-Beispiele sind in der Beispiel-Datenbank als Bluesky-Posts hinterlegt (Cross-Posts der offiziellen Grünen-Accounts). Orientiere dich an Ton, Aufbau und Formatierung.`,
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit',
-    title: 'LinkedIn',
-    description: 'LinkedIn-Posts & Artikel',
-    iconKey: 'PiLinkedinLogo',
-    avatar: '💼',
-    backgroundColor: '#316049',
-    mention: 'linkedin',
-    contextPrefix: '[Plattform: LinkedIn]',
-    skillCategory: 'social',
-    promptTemplate: 'LinkedIn-Post zu: ',
-    skillSystemPrompt: `**LINKEDIN-POST (max. 600 Zeichen):**
-
-Professionell aber zugänglich, Analysen und Einblicke, minimale Emojis. Adressiere Multiplikator*innen und Fachpublikum.
-
-**Beispiel-Suche:** Nutze IMMER \`search_examples\` mit \`platform="linkedin"\`, falls Beispiele vorhanden sind. Sonst orientiere dich an einem ruhigen, analytischen Ton.`,
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit',
-    title: 'Reel / TikTok',
-    description: 'Reel- & TikTok-Skripte',
-    iconKey: 'PiTiktokLogo',
-    avatar: '🎬',
-    backgroundColor: '#316049',
-    mention: 'reel',
-    contextPrefix: '[Plattform: Reel/TikTok-Skript]',
-    skillCategory: 'social',
-    promptTemplate: 'Skript zu folgendem Thema: ',
-    skillSystemPrompt: `**REEL / TIKTOK-SKRIPT (max. 1500 Zeichen):**
-
-Skript-Format mit klarer Zeitstruktur:
-- 00:00–00:20 **Hook** (direkter Aufhänger, der scrollt-stoppt)
-- 00:20–01:10 **Main** (Kernbotschaft, eine Aussage)
-- 01:10–01:30 **CTA** (Handlungsaufforderung)
-
-Schreibe gesprochene Sprache, kurze Sätze, visuelle Cues in eckigen Klammern \`[Schnitt: …]\`.`,
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit',
-    title: 'Aktionsideen',
-    description: 'Kreative Aktionsideen entwickeln',
-    iconKey: 'PiLightbulb',
-    avatar: '💡',
-    backgroundColor: '#316049',
-    mention: 'aktion',
-    contextPrefix: '[Plattform: Aktionsideen]',
-    skillCategory: 'sonstiges',
-    promptTemplate: 'Entwickle Aktionsideen zu: ',
-  },
-  {
-    identifier: 'gruenerator-rede-schreiber',
-    title: 'Rede',
-    description: 'Politische Reden',
-    iconKey: 'PiMicrophoneStage',
-    avatar: '🎙️',
-    backgroundColor: '#316049',
-    mention: 'rede',
-    skillCategory: 'presse',
-    promptTemplate: 'Schreibe eine Rede zum Thema: ',
-  },
-  {
-    identifier: 'gruenerator-wahlprogramm',
-    title: 'Wahlprogramm',
-    description: 'Programmkapitel',
-    iconKey: 'PiListChecks',
-    avatar: '📋',
-    backgroundColor: '#316049',
-    mention: 'wahlprogramm',
-    skillCategory: 'dokumente',
-  },
-  {
-    identifier: 'gruenerator-leichte-sprache',
-    title: 'Leichte Sprache',
-    description: 'Texte in Leichte Sprache übersetzen',
-    iconKey: 'PiTranslate',
-    avatar: '🗣️',
-    backgroundColor: '#316049',
-    mention: 'leichte-sprache',
-    skillCategory: 'sonstiges',
-    promptTemplate: 'Übersetze folgenden Text in Leichte Sprache: ',
-  },
-  // ─── Per-Landesverband shortcuts ───
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit-berlin',
-    title: 'PM Berlin',
-    description: 'Pressemitteilung im Stil Grüne Berlin',
-    iconKey: 'PiBuildings',
-    avatar: '🐻',
-    backgroundColor: '#316049',
-    mention: 'presse-berlin',
-    contextPrefix: '[LV: Berlin / Modus: PM]',
-    skillCategory: 'presse',
-    promptTemplate: 'Schreibe eine Pressemitteilung im Stil Grüne Berlin zum Thema: ',
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit-berlin',
-    title: 'Social Berlin',
-    description: 'Social-Media-Post im Stil Grüne Berlin',
-    iconKey: 'PiBuildings',
-    avatar: '🐻',
-    backgroundColor: '#316049',
-    mention: 'social-berlin',
-    contextPrefix: '[LV: Berlin / Modus: Social Media]',
-    skillCategory: 'social',
-    promptTemplate:
-      'Erstelle Social-Media-Posts im Stil Grüne Berlin (Facebook, Instagram, Twitter/X, LinkedIn) zum Thema: ',
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit-hamburg',
-    title: 'PM Hamburg',
-    description: 'Pressemitteilung im Stil Grüne Hamburg',
-    iconKey: 'PiAnchor',
-    avatar: '⚓',
-    backgroundColor: '#316049',
-    mention: 'presse-hamburg',
-    contextPrefix: '[LV: Hamburg / Modus: PM]',
-    skillCategory: 'presse',
-    promptTemplate: 'Schreibe eine Pressemitteilung im Stil Grüne Hamburg zum Thema: ',
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit-hamburg',
-    title: 'Social Hamburg',
-    description: 'Social-Media-Post im Stil Grüne Hamburg',
-    iconKey: 'PiAnchor',
-    avatar: '⚓',
-    backgroundColor: '#316049',
-    mention: 'social-hamburg',
-    contextPrefix: '[LV: Hamburg / Modus: Social Media]',
-    skillCategory: 'social',
-    promptTemplate: 'Erstelle Social-Media-Posts im Stil Grüne Hamburg zum Thema: ',
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit-mecklenburg-vorpommern',
-    title: 'PM MV',
-    description: 'Pressemitteilung im Stil Grüne Mecklenburg-Vorpommern',
-    iconKey: 'PiWaves',
-    avatar: '🌊',
-    backgroundColor: '#316049',
-    mention: 'presse-mv',
-    contextPrefix: '[LV: Mecklenburg-Vorpommern / Modus: PM]',
-    skillCategory: 'presse',
-    promptTemplate: 'Schreibe eine Pressemitteilung im Stil Grüne MV zum Thema: ',
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit-mecklenburg-vorpommern',
-    title: 'Social MV',
-    description: 'Social-Media-Post im Stil Grüne Mecklenburg-Vorpommern',
-    iconKey: 'PiWaves',
-    avatar: '🌊',
-    backgroundColor: '#316049',
-    mention: 'social-mv',
-    contextPrefix: '[LV: Mecklenburg-Vorpommern / Modus: Social Media]',
-    skillCategory: 'social',
-    promptTemplate: 'Erstelle Social-Media-Posts im Stil Grüne MV zum Thema: ',
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit-thueringen',
-    title: 'PM Thüringen',
-    description: 'Pressemitteilung im Stil Bündnisgrüne Thüringen',
-    iconKey: 'PiTree',
-    avatar: '🌲',
-    backgroundColor: '#316049',
-    mention: 'presse-thueringen',
-    contextPrefix: '[LV: Thüringen / Modus: PM]',
-    skillCategory: 'presse',
-    promptTemplate: 'Schreibe eine Pressemitteilung im Stil Bündnisgrüne Thüringen zum Thema: ',
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit-thueringen',
-    title: 'Social Thüringen',
-    description: 'Social-Media-Post im Stil Bündnisgrüne Thüringen',
-    iconKey: 'PiTree',
-    avatar: '🌲',
-    backgroundColor: '#316049',
-    mention: 'social-thueringen',
-    contextPrefix: '[LV: Thüringen / Modus: Social Media]',
-    skillCategory: 'social',
-    promptTemplate: 'Erstelle Social-Media-Posts im Stil Bündnisgrüne Thüringen zum Thema: ',
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit-brandenburg',
-    title: 'PM Brandenburg',
-    description: 'Pressemitteilung im Stil Brandenburger Bündnisgrüne',
-    iconKey: 'PiFlowerLight',
-    avatar: '🌻',
-    backgroundColor: '#316049',
-    mention: 'presse-brandenburg',
-    contextPrefix: '[LV: Brandenburg / Modus: PM]',
-    skillCategory: 'presse',
-    promptTemplate: 'Schreibe eine Pressemitteilung im Stil Brandenburger Bündnisgrüne zum Thema: ',
-  },
-  {
-    identifier: 'gruenerator-oeffentlichkeitsarbeit-brandenburg',
-    title: 'Social Brandenburg',
-    description: 'Social-Media-Post im Stil Brandenburger Bündnisgrüne',
-    iconKey: 'PiFlowerLight',
-    avatar: '🌻',
-    backgroundColor: '#316049',
-    mention: 'social-brandenburg',
-    contextPrefix: '[LV: Brandenburg / Modus: Social Media]',
-    skillCategory: 'social',
-    promptTemplate: 'Erstelle Social-Media-Posts im Stil Brandenburger Bündnisgrüne zum Thema: ',
-  },
-] as const satisfies readonly SystemSkill[];
-
-const mentionMap = new Map<string, string>(
-  SKILLS.map((skill) => [skill.mention.toLowerCase(), skill.identifier])
-);
-
-export function resolveSkillMention(alias: string): string | null {
-  return mentionMap.get(alias.toLowerCase()) ?? null;
-}
