+++ packages/shared/src/agents/skills/kommunalpolitik.ts
@@ -1,0 +1,13 @@
+import type { SystemSkill } from './types.js';
+
+export const KOMMUNALPOLITIK_SKILL = {
+  identifier: 'gruenerator-antrag',
+  title: 'Kommunalpolitik',
+  description: 'Haushalt bewerten, beraten, Anträge & Resolutionen entwerfen',
+  iconKey: 'PiBuildings',
+  avatar: '🏛️',
+  backgroundColor: '#316049',
+  mention: 'kommunal',
+  skillCategory: 'dokumente',
+  isSystemDefault: true,
+} as const satisfies SystemSkill;
