+++ packages/shared/src/agents/skills/types.ts
@@ -1,0 +1,4 @@
+import type { SystemAgentId } from '../system.js';
+import type { Skill } from '../types.js';
+
+export type SystemSkill = Skill & { identifier: SystemAgentId };
